+++ XADMaster/CSHandle.h
@@ -36,7 +36,7 @@ extern NSString *CSNotSupportedException;
 -(void)seekToFileOffset:(off_t)offs;
 -(void)seekToEndOfFile;
 -(void)pushBackByte:(int)byte;
--(int)readAtMost:(off_t)num toBuffer:(void *)buffer;
+-(int)readAtMost:(int)num toBuffer:(void *)buffer;
 -(void)writeBytes:(int)num fromBuffer:(const void *)buffer;
