+++ XADMaster/CSMemoryHandle.m
@@ -75,7 +75,7 @@
 
 //-(void)pushBackByte:(int)byte {}
 
--(int)readAtMost:(off_t)num toBuffer:(void *)buffer
+-(int)readAtMost:(int)num toBuffer:(void *)buffer
 {
 	if(!num) return 0;
