+++ XADMaster/CSMultiHandle.h
@@ -25,7 +25,7 @@ extern NSString *CSSizeOfSegmentUnknownException;
 
 -(void)seekToFileOffset:(off_t)offs;
 -(void)seekToEndOfFile;
--(int)readAtMost:(off_t)num toBuffer:(void *)buffer;
+-(int)readAtMost:(int)num toBuffer:(void *)buffer;
 
 -(void)_raiseSizeUnknownForSegment:(int)i;
