+++ XADMaster/CSSubHandle.h
@@ -18,6 +18,6 @@
 
 -(void)seekToFileOffset:(off_t)offs;
 -(void)seekToEndOfFile;
--(int)readAtMost:(off_t)num toBuffer:(void *)buffer;
+-(int)readAtMost:(int)num toBuffer:(void *)buffer;
 
 @end
