+++ XADMaster/XADXORHandle.h
@@ -17,6 +17,6 @@
 -(BOOL)atEndOfFile;
 -(void)seekToFileOffset:(off_t)offs;
 -(void)seekToEndOfFile;
--(int)readAtMost:(off_t)num toBuffer:(void *)buffer;
+-(int)readAtMost:(int)num toBuffer:(void *)buffer;
 
 @end
